GroupHelper created.
