+++ fixture/application.py
@@ -1,6 +1,7 @@
 from selenium.webdriver.firefox.webdriver import WebDriver
 from selenium.webdriver.support.ui import Select
 from fixture.session import SessionHelper
+from fixture.group import GroupHelper
 
 
 class Application:
@@ -9,6 +10,7 @@ class Application:
         self.wd = WebDriver()
         self.wd.implicitly_wait(60)
         self.session = SessionHelper(self)
+        self.group = GroupHelper(self)
 
     def open_home_page(self):
         wd = self.wd
@@ -17,31 +19,6 @@ class Application:
     def return_home(self):
         wd = self.wd
         wd.find_element_by_link_text("home page").click()
-
-    def open_groups_page(self):
-        wd = self.wd
-        wd.find_element_by_link_text("groups").click()
-
-    def create_group(self, group):
-        wd = self.wd
-        # open groups page
-        self.open_groups_page()
-        # open group creation form
-        wd.find_element_by_name("new").click()
-        # fill group form
-        wd.find_element_by_name("group_name").click()
-        wd.find_element_by_name("group_name").clear()
-        wd.find_element_by_name("group_name").send_keys(group.name)
-        wd.find_element_by_name("group_header").click()
-        wd.find_element_by_name("group_header").clear()
-        wd.find_element_by_name("group_header").send_keys(group.header)
-        wd.find_element_by_name("group_footer").click()
-        wd.find_element_by_name("group_footer").clear()
-        wd.find_element_by_name("group_footer").send_keys(group.footer)
-        # submit group creation
-        wd.find_element_by_name("submit").click()
-        # check that group was created
-        self.open_groups_page()
 
     def create_contact(self, contact):
         wd = self.wd
