+++ test/test_add_group.py
@@ -15,12 +15,12 @@ def app(request):
 def test_add_group(app):
     app.open_home_page()
     app.session.login(username="admin", password="secret")
-    app.create_group(Group(name="group_002", header="another_header", footer="another_footer"))
+    app.group.create(Group(name="group_002", header="another_header", footer="another_footer"))
     app.session.logout()
 
 
 def test_add_empty_group(app):
     app.open_home_page()
     app.session.login(username="admin", password="secret")
-    app.create_group(Group(name="", header="", footer=""))
+    app.group.create(Group(name="", header="", footer=""))
     app.session.logout()
